Refactor AppSidebar component structure and update import paths
- Moved AppSidebar component to a new directory structure for better organization.
- Removed SidebarAction component as it is no longer needed.
- Updated import paths in layout.tsx to reflect the new location of AppSidebar.
