+++ components/AppSidebar/SidebarAction.tsx
@@ -8,10 +8,10 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@radix-ui/react-dropdown-menu";
-import { SidebarMenuButton } from "../sidebar";
 import { ChevronUp } from "lucide-react";
 import { redirect } from "next/navigation";
 import { User } from "@supabase/supabase-js";
+import { SidebarMenuButton } from "../ui/sidebar";
 const SidebarAction = ({ user }: { user: User }) => {
   return (
     <DropdownMenu>
